Add azure pipelines definition for .net standard 2.0 libraries
The pipeline definition builds on Windows Server 2019 and runs unit tests on the following operating systems and target frameworks:
- Windows Server 2019, .NET 4.6.1      (net461)
- Windows Server 2019, .NET Core 2.1   (netcoreapp2.1)
- Windows Server 2019, .NET Core 3.1   (netcoreapp3.1)
- Windows Server 2019, .NET 5.0        (net5.0, no platform specific TPMs)
- Ubuntu 20.04,        .NET Core 2.1   (netcoreapp2.1)
- Ubuntu 20.04,        .NET Core 3.1   (netcoreapp3.1)
- Ubuntu 20.04,        .NET 5.0        (net5.0, no platform specific TPMs)

After tests have completed successfully nuget packages are created according to their specification (*.nuspec).
If the build was tagged with a tag starting with 'v' the nuget packages are pushed to nuget.org.